requirements.txt
# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# LLM and AI
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
litellm>=1.20.0
openai>=1.10.0
google-generativeai>=0.4.0  # Google AI embeddings (FREE tier)

# Vector database
qdrant-client>=1.7.0

# Twilio for WhatsApp
twilio>=8.10.0

# Google Cloud
google-cloud-firestore>=2.14.0
google-cloud-pubsub>=2.19.0

# AWS for sentiment
boto3>=1.34.0

# Utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
httpx>=0.26.0
tenacity>=8.2.0
structlog>=24.1.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
